backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
google-genai
pypdf
pillow
python-multipart